requirements.txt
chromadb==0.5.5
openai>=1.40.0
pydantic>=2.7.0
python-dotenv>=1.0.1
tiktoken>=0.7.0
Typer>=0.12.0
uvicorn>=0.30.0
fastapi>=0.111.0
ragas>=0.1.14
pytest>=8.2.0
streamlit==1.37.0
streamlit-mic-recorder==0.0.4
